requirements.txt
requests
pycryptodome
bs4
lxml
pandas
pymysql
phpserialize